Move constraints_test_utils.hh into gcs/constraints/innards/
The header is only used by *_test.cc files (and a couple of examples
that lift the test machinery for ad-hoc verification), and was already
excluded from the public install. Moving it to gcs/constraints/innards/
puts it next to reified_dispatcher.hh as another constraint-internal
header, and signals more clearly that it isn't part of the user-facing
API.

Updates the include guard for the new path, updates 24 *_test.cc
files, the n_fractions example, and resorts five include lines that
the path change put out of alphabetical order.

The CMake install REGEX still excludes constraints_test_utils.hh from
the public install — it's anchored on the filename, so the path
change doesn't affect it.
