+++ gcs/constraints/all_different_test.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/all_different.hh>
-#include <gcs/constraints/constraints_test_utils.hh>
+#include <gcs/constraints/innards/constraints_test_utils.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
