+++ gcs/constraints/equals_test.cc
@@ -1,5 +1,5 @@
-#include <gcs/constraints/constraints_test_utils.hh>
 #include <gcs/constraints/equals.hh>
+#include <gcs/constraints/innards/constraints_test_utils.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
