+++ gcs/constraints/innards/constraints_test_utils.hh
@@ -1,5 +1,5 @@
-#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_CONSTRAINTS_TEST_UTILS_HH
-#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_CONSTRAINTS_TEST_UTILS_HH
+#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_INNARDS_CONSTRAINTS_TEST_UTILS_HH
+#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_INNARDS_CONSTRAINTS_TEST_UTILS_HH
 
 #include <gcs/current_state.hh>
 #include <gcs/exception.hh>
