+++ gcs/constraints/lex_test.cc
@@ -1,4 +1,4 @@
-#include <gcs/constraints/constraints_test_utils.hh>
+#include <gcs/constraints/innards/constraints_test_utils.hh>
 #include <gcs/constraints/lex.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
